test: move insecure HTTP CLI tests
